Set up quiz routes
