+++ src/api/routes/index.js
@@ -1,0 +1,4 @@
+import questionRouter from './question/question.router';
+import quizRouter from './quiz/quiz.router';
+
+export const routes = { questionRouter, quizRouter };
