--- src/routes/index.js
@@ -1,3 +1,0 @@
-import router from './base/nav';
-
-export default router;
